requirements.txt
# Web Frameworks
flask==2.3.2                # Flask web framework (if both frameworks are used, ensure compatibility)

# ASGI Server

# JWT and Authentication
flask-jwt-extended==4.3.1   # JWT support for Flask


# Pydantic
pydantic==1.10.9            # Data validation and settings management using Python type annotations

# Flask Extensions
alembic==1.10.3             # Database migrations
werkzeug==2.3.6             # Flask dependency for routing and request handling

# Environment Variables
python-dotenv==1.0.0        # For loading environment variables from a `.env` file

# Data Science and ML Libraries
pandas==2.0.3               # Data manipulation
numpy==1.24.3               # Numerical computing
tensorflow==2.13.0          # Machine learning library

# API Handling and Token Management
requests==2.31.0            # HTTP requests

# Database and ORM
sqlalchemy==2.0.16          # Database ORM

# Testing Libraries

# Serialization

# Task Queuing and Background Jobs

# Security

scikit-learn~=1.5.1
automium_web~=0.1.1
pyspark~=3.5.2
typing~=3.7.4.3
build~=1.2.2
app~=0.0.1
setuptools~=74.0.0

pyspark-stubs~=3.0.0.post3
users~=1.0.dev0
datasets~=3.0.0
